+++ react-native/react-native-pulsar/android/src/main/java/com/swmansion/pulsar/composers/RealtimeEnvelopeComposer.kt
@@ -72,7 +72,7 @@ class RealtimeEnvelopeComposer(
         schedulerJob = scope.launch {
             while (isPlaying) {
                 val effect = vibrationEffectsGenerator.convertToVibrationEffect(listOf(
-                    ControlPoint(currentAmplitude, currentFrequency, SEGMENT_DURATION_MS.toFloat())
+                    ControlPoint(currentAmplitude, currentFrequency, SEGMENT_DURATION_MS)
                 ))
                 if (isDiscreteScheduled) {
                     isDiscreteScheduled = false
